% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{submit_ENSC311}
\alias{submit_ENSC311}
\title{Create a homework submission file for ENSC311 tutorials}
\usage{
submit_ENSC311(tutorial_index = NULL)
}
\arguments{
\item{tutorial_idx}{An index indicating which homework submission file to
retrieve.  If NULL, the vector of available homework submission files is
returned.}
}
\value{
Either a vector of available homework submission files or TRUE if the
  homework submission file was created in your working directory correctly.
}
\description{
Creates a homework submission file in your working directory.
}
\details{
To use this function, first run the function without passing arguments.
Examine the resulting list of files to be created and then run the function
again, passing the index number of the file to be created.

You can determine your working directory (where the file is created) with
\code{\link{getwd}}.
}
