% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_ENSC311}
\alias{update_ENSC311}
\title{Update the ENSC311 package}
\usage{
update_ENSC311(instructor)
}
\description{
Automatically installs the latest version of ENSC311 from github.
}
\details{
Simply run "update_ENSC311()" to get the latest version of the package.
}
